imx8mm-var-dart: configure op-tee memory to 2gb
For now, dram size is hardcoded in optee-os. Since the DART-MX8M-MINI and
VAR-SOM-MX8M-MINI evks ship with 2GB by default, use 2GB as the default
configuration.

In the future, it should be investigated if this can be set dynamically
from the EEPROM on all Variscite SoMs.

This patch is derived from: 01e70a1

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>