linux-firmware: bc43xx: add symbolic link to firmware for each model
The Linux driver requests a firmware named according the machine compatible
string reported in the device tree.
For example, for DART-MX8M-PLUS the machine compatible string in the device
tree file is <compatible = "variscite,imx8mp-var-dart", "fsl,imx8mp">
so the driver tries loading firmware files with the following name:
- brcmfmac43430-sdio.variscite,imx8mp-var-dart.bin
- brcmfmac43430-sdio.variscite,imx8mp-var-dart.txt
If the files are missing the driver calls the fallback function that searches
for the native name of firmware files:
- brcmfmac43430-sdio.bin
- brcmfmac43430-sdio.txt
In this case we see the following kernel warnings:
- brcmfmac mmc0:0001:1: Direct firmware load for brcm/brcmfmac4339-sdio.variscite,imx8mp-var-dart.bin failed with error -2
- brcmfmac mmc0:0001:1: Direct firmware load for brcm/brcmfmac4339-sdio.variscite,imx8mp-var-dart.txt failed with error -2
The changes in this patch create symbolic links to the expected files.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>