freertos-variscite: Replace MX93 toolchain with cortexm-toolchain-cross.
Utilize the new dedicated toolchain package cortexm-toolchain-cross. In
effect, this uses the same underlying toolchain, but in this case, it is
no longer retrieved via freertos-variscite. This is more portable and
resolves incompatibilities found using devtool (class externalsrc).

In the future, this functionality can be unified with other platforms,
but only apply to the MX93 for now.

Signed-off-by: Ken Sloat <ken.s@variscite.com>