u-boot-variscite: Set PV explicitly to something meaningful.
Currently in the U-Boot recipe, PV is unset. This results in a defaulted
value of 1.0. This is not particularly useful and for informational
purposes should be set explicitly.

To correct this, use the U-Boot branch name plus the git sha (7 chars).
This results in a PV that looks like "lf_v2022.04_var02+git277ff23"

Signed-off-by: Ken Sloat <ken.s@variscite.com>