cortexm-toolchain-cross: Fix incorrect license file checksum
The current recipe has a copy/paste error for LIC_FILES_CHKSUM. This has
gone unoticed because to date, any recipes that DEPEND on
cortexm-toolchain-cross have no installed package in any image. Such
recipes will not trigger a QA error if simply built, however if
installed, the build will fail to complete.

To fix this, simply set the appropriate license, relying on the existing
BSD-3-Clause file contained in the COMMON_LICENSE_DIR.

Signed-off-by: Ken Sloat <ken.s@variscite.com>