bcm43xx-utils: imx93-var-som: Work around Wi-Fi reboot bug.
When performing an immediate reboot following a startup, it is possible
that such a reboot call, when made by another task, would still allow
for the variscite-wifi service to startup and then subsequently stop.
When this situation occurs, it sometimes results in the cfg80211
driver's regulatory framework making a call that results in a memory
abort (level 3 translation fault). This appears to be a generic bug in
the kernel workqueue or timer framework - likely a "use-after-free" race
condition.

In debugging and testing, I found that this bug can be avoided
altogether by adding a short delay to the system before proceeding with
reboot. Following this delay, if the system is in a reboot cycle, the
routine should proceed normally. Thus, introduce this delay in the Wi-Fi
startup function to work around the problem for now.

Signed-off-by: Ken Sloat <ken.s@variscite.com>