variscite: Add systemd override conf file
The poky distro config in Langdale makes a default hard asignment of
INIT_MANAGER = "sysvinit." Since NXP and the community inherit the poky
distro, if using their distros, it is not possible to override this
variable in other layers. While it is still possible to enable systemd
via individual variables, this behavior will result in a distro
containing both systemd and sysvinit features.

With certain packages (like networkmanager), this can result in
duplication between systemd and sysvinit scripts and thus cause runtime
conflicts.

While we cannot set INIT_MANAGER directly (even via an override) without
patching poky or creating our own distro, we can undo the effects of it.
Thus, create a conf include file which can be included by machines
wishing to invoke this behavior. This can then be utilized by the
desired machine by simply appending an OVERRIDE for "broken_initmanager."

See the initial background discussion of this poky behavior that was
introduced with Langdale here:
https://lists.openembedded.org/g/openembedded-core/topic/92210454

and subsequent patch that corrected it in Mickledore:
https://git.yoctoproject.org/poky/commit/meta-poky/conf/distro/poky.conf?h=mickledore&id=316d66b4c43a2338609ca6c1c51afb0970a41c8e

Signed-off-by: Ken Sloat <ken.s@variscite.com>