pm-utils: disable stop and start for wifi and bluetooth
Using DART-6UL with NAND flash, when the BT_EN_GPIO is set to 0 before going
to suspend, then upon wakeup the UBI filesystem is getting corrupted and is being
set to read-only.

To avoid this issue, when booting with NAND controller enabled, the WiFi and Bluetooth
stop/start are disabled for both DART-6UL and VAR-SOM-6UL SOMs.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>