freertos-variscite: force compilation to respect PARALLEL_MAKE
The compilation of freertos-variscite's applications is done by
running the build_all.sh script, that calls "make -j" to compile
the application. When the argument "-j" is used by make, without
specifying any value, there is no limit to the number of parallel
jobs used during compilation. This was causing the freertos-variscite
to consume all the threads available during compilation, even if the
PARALLEL_MAKE variable was being defined.

This patch overrides the "-j" argument in build_all.sh to respect the
value defined in PARALLEL_MAKE.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>