freertos-variscite: imx8mp-var-dart: Skip install of any missing targets
A bug was found while configuring for all-demos where some demos (source
code) actually omit certain targets (i.e. ddr_debug for rdc). This will
result in executing a bogus install command which may fail and halt the
do_install task. To fix this, simply check if the expected target
directory even exists, and if not, skip that particular target.

Signed-off-by: Ken Sloat <ken.s@variscite.com>