imx93-var-som: iw612-bt: fix stop function if hci0 is down
When disabling bluetooth, it is necessary to restore the baud to 115200
using hcitool so that it is in the expected state next time it is started.

hcitool requires that the hci interface is up.

Thus, this patch checks if hci0 is down and brings it up if necessary
before running the hcitool command.

Signed-off-by: Nate Drude <nate.d@variscite.com>