recipes-core: move base-files recipe logic to udev-extraconf
As the base-files bbappend meanwhile only holds logic for modprobe
configurations, it's suitable to move it to udev-extraconf, being
a more related place for it.

This will also help to add those extra modprobe configurations to
Debian, as the base-files package is provided by debian-base masking
the base-files package from Yocto.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>