Revert "base-files: Add variscite-hdmi-audio.conf to /etc/modprobe.d"
This reverts commit dfd9f250abcd1d180896d5381d9793d1ca9603a1.

Managing the order of snd-soc-wm8904 and snd-soc-hdmi-codec is not
a complete solution to force WM8904 as the default ALSA device. It does
not consider other sound cards present in the system.

For example, the DART-MX8M-PLUS has three sound cards:

root@imx8mp-var-dart:~# cat /proc/asound/cards
 0 [audiohdmi      ]: audio-hdmi - audio-hdmi
                      audio-hdmi
 1 [imxaudioxcvr   ]: imx-audio-xcvr - imx-audio-xcvr
                      imx-audio-xcvr
 2 [wm8904audio    ]: simple-card - wm8904-audio
                      wm8904-audio

The correct way to set WM8904 as the default card is by configuring
/etc/asound.conf. A patch for this will follow.

Signed-off-by: Nate Drude <nate.d@variscite.com>