[RND-2326] linux-variscite: fix duplicated kernel version name
We have previously seen weird kernel module directory names in which the
LOCALVERSION from the kernel recipe was duplicated, e.g.
6.6.23-imx8mp-imx8mp-g6cb72986bc16

Fix it and use the standardized way that is also used by kernels
directly from the Yocto project.

The problem with the LOCALVERSION approach comes from doing magic in
https://github.com/Freescale/meta-freescale/blob/master/classes/fsl-kernel-localversion.bbclass#L16
(which is also used by us as it is a class we inherit), but which is
later reverted by NXP
(https://github.com/nxp-imx/meta-imx/blob/scarthgap-6.6.36-2.1.0/meta-imx-bsp/recipes-kernel/linux/linux-imx_6.6.bb#L58)
which is not used by us.

Also, unify the version string to '-var-lts-next' instead of having
individual version strings for each machine.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>