imx-isp: Prevent unnecessary service restarts for exit status 6
The imx8-isp service exits with status 6 when a device tree lacks
Basler support. Since this failure is expected in such cases,
repeated restart attempts are unnecessary and undesirable.

This patch updates the systemd service file to prevent automatic
restarts when the service exits with status 6, reducing redundant
logs and resource usage.

Signed-off-by: Nate Drude <nate.d@variscite.com>