freertos-variscite: variscite-rproc-common: Rework dtb selection
The query of the SOM revision will fail for i.MX93-based SOMs, as they
do have the EEPROM on i2c bus address 2, while the bus address is hard-
coded to 0.

As this can be also the case for other supported SOMs, fix it by limit
the dtb selection to those who need it.

Fixes:
- https://github.com/varigit/meta-variscite-bsp-imx/commit/38d8b9b411cb3b324865198dff46d1763f90b2f0
- https://github.com/varigit/meta-variscite-bsp-imx/commit/84c7ab48be37ecbb2a56527d76afad41f1157592

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>