alsa-state: provide default mixer settings for wm8904 on mx6ul
Provide default mixer settings for wm8904 on mx6ul to ensure that the
audio device is not muted as a default setting.

The default audio volume (and it not being muted) was taken from them
imx-nxp-bsp folder which already uses the wm8904, the rest was taken
from the default settings read back after the start.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>