pm-utils-variscite: Fix Broadcom WiFi driver unloading during suspend
On 6UL SOMs with a Broadcom WiFi module, the WiFi driver should only be
unloaded during suspend if the `sdio,must-keep-power-in-suspend` property
is NOT present in the device tree. The previous logic relied on checking
for a NAND controller, which was too restrictive and unreliable for
carrier board like Symphony. This update ensures the driver is correctly
reloaded after power is restored following suspend/resume.

Signed-off-by: Marius Bagassien <marius.b@variscite.com>