machine.conf: mx6ul: bbmask psplash recipes from meta-imx
There are changes in the meta-imx that are (probably) good for later imx
platforms, but have a negative impact on platforms with lower horse
power:

- Detection of framebuffer device takes extra time to check
- Necessity to wait for udev services to complete mean that the start of
psplash is delayed significantly
- Progress is shown so late that we are close to the X11 startup anyway

All of this is fixed if we simply stick to the vanilla implementation
provided by the poky. That way psplash progress bar is displayed earlier
during the boot and we also have a proper progress indication.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>